requirements.txt
requests
tinycss
attrs
